+++ app/src/main/java/com/brewthings/app/util/ComposeUtils.kt
@@ -1,0 +1,21 @@
+package com.brewthings.app.util
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+
+@Composable
+fun <T : Any?> newOrCached(
+    data: T,
+    initialValue: T,
+): T {
+    var previousData: T by remember { mutableStateOf(initialValue) }
+    return if (data != null) {
+        previousData = data
+        data
+    } else {
+        previousData
+    }
+}
